% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Simulate one dataset for a given (pi0, target_power, m, n) scenario.}
\usage{
simulate_dataset(pi0, target_power, m, n, effect_grid)
}
\arguments{
\item{pi0}{True proportion of nulls}

\item{target_power}{Target \emph{average} power used for calibration}

\item{m}{Total number of hypotheses}

\item{n}{Sample size per group}

\item{effect_grid}{Output of build_effect_grid()}
}
\value{
A list with:
\itemize{
\item pvals   - numeric vector of p-values (length m)
\item tstats  - numeric vector of t-statistics (length m)
\item is_null - logical vector of length m, TRUE if hypothesis is null
}
}
\description{
Simulate one dataset for a given (pi0, target_power, m, n) scenario.
}
